Basic neat is working. It takes in an input file, or uses H1N1.fa by default, and outputs a single fastq of shuffled reads

+++ src/utils/cli.rs
@@ -1,7 +1,7 @@
 use std::ffi::CString;
 use clap::Parser;
 
-#[derive(Parser)]
+#[derive(Parser, Debug)]
 pub struct Cli {
     #[arg(short='y', long="configuration_yaml", default_value_t=String::new())]
     pub config: String,
@@ -12,7 +12,7 @@ pub struct Cli {
     #[arg(short='f', long="output_file_prefix", default_value_t=String::new())]
     pub output_file_prefix: String,
     #[arg(short='l', long="read_len", default_value_t = 150)]
-    pub read_length: u32,
+    pub read_length: usize,
     #[arg(short='c', long="coverage", default_value_t = 10)]
-    pub coverage: u32,
+    pub coverage: usize,
 }
